FY-108/feat: 홈이 아닌 별도의 콜백 페이지에서 액세스 토큰 발급 처리
