+++ src/App.js
@@ -9,6 +9,7 @@ import HomePage from "./pages/HomePage";
 import GamePlayPage from "./pages/game/GamePlayPage";
 import FriendsPage from "./pages/FriendsPage";
 import LoginPage from "./pages/LoginPage";
+import LoginCallbackPage from "./pages/LoginCallbackPage";
 import UserProfilePage from "./pages/UserProfilePage";
 import MyProfilePage from "./pages/MyProfilePage";
 import RoomListPage from "./pages/RoomListPage";
@@ -33,6 +34,8 @@ const AppContent = () => {
       <Routes>
         <Route path="/" element={<HomePage />} />
         <Route path="/login" element={<LoginPage />} />
+        {/* 로그인 콜백 페이지를 라우팅 관련 이슈 없이 /login 경로 아래 둘 수 있는 방법 알아보기 */}
+        <Route path="/callback" element={<LoginCallbackPage />} />
         <Route path="/profile" element={<MyProfilePage />} />
         <Route path="/profile/users/:user_id" element={<UserProfilePage />} />
         <Route path="/users" element={<UsersPage />} />
